Full refactor of keyboard file structure
