+++ keyboards/handwired/do43/config.h
@@ -48,7 +48,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define DIODE_DIRECTION COL2ROW
 
 /* Custom OLED font */
-#define OLED_FONT_H "keyboards/handwired/do43/keymaps/roccojiang/glcdfont.c"
+#define OLED_FONT_H "keyboards/handwired/do43/custom_font.h"
 
 /* QMK DFU Bootloader */
 #define QMK_ESC_OUTPUT C6  // COL pin if COL2ROW
