--- keyboards/handwired/do43/keymaps/roccojiang/oled.c
@@ -1,349 +1,0 @@
-/* Copyright 2021 Rocco Jiang
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include QMK_KEYBOARD_H
-
-/*
- * Bongo Cat WPM visualiser.
- * 
- * Credits:
- * @j-inc for initial bongocat code
- * @obosob for initial animation approach
- * @pixelbenny for images
- * @aseiger for initial adaptation to 128x32 display
- * @vectorstorm for initial run length encoding to compress animation size
- * @filterpaper for improvements to RLE
- */
-
-#define IDLE_FRAMES 5
-#define TAP_FRAMES 2
-
-#define IDLE_SPEED 50 // animation idles below this wpm value
-#define TAP_SPEED 75 // animation triggers above this wpm value
-
-#define FRAME_DURATION 200 // ms per frame
-
-uint32_t anim_timer = 0;
-uint32_t anim_sleep = 0;
-
-// Bongocat frames are encoded in pairs of 'byte, count'
-// RLE decode will read each pair and render bytes based on count
-static void decode_frame(unsigned char const *frame) {
-	uint16_t cursor = 0;
-	uint8_t size = pgm_read_byte_near(frame);
-
-	oled_set_cursor(0, 0);
-	for (uint8_t i = 1; i < size - 1; i += 2) {
-		// Loop through pairs of byte and count
-		char byte = pgm_read_byte_near(frame + i);
-		uint8_t count = pgm_read_byte_near(frame + i + 1);
-
-		for (uint8_t j = 0; j < count; ++j) {
-			// Render byte repetition
-			oled_write_raw_byte(byte, cursor);
-			cursor++;
-		}
-	}
-}
-
-static void render_cat_idle(void) {
-    // Run-length encoded idle frames
-	static unsigned char const idle0[] PROGMEM = {181,
-		0x00,0x33,0x80,0x02,0x00,0x05,0x80,0x01,0x40,0x02,0x20,0x04,0x10,0x04,0x08,0x01,
-		0x04,0x01,0x02,0x01,0x01,0x02,0x02,0x01,0x0c,0x01,0x30,0x01,0x40,0x01,0x80,0x01,
-		0x00,0x2e,0x80,0x03,0x00,0x31,0x1e,0x01,0xe1,0x01,0x00,0x02,0x01,0x02,0x02,0x02,
-		0x01,0x01,0x00,0x04,0x80,0x02,0x00,0x09,0x80,0x01,0x00,0x01,0x30,0x02,0x00,0x01,
-		0xc0,0x01,0xc1,0x02,0xc2,0x01,0x04,0x01,0x08,0x01,0x10,0x01,0x20,0x01,0x40,0x01,
-		0x80,0x01,0x00,0x03,0x80,0x04,0x40,0x04,0x20,0x04,0x10,0x04,0x08,0x05,0x04,0x05,
-		0x02,0x04,0x01,0x04,0x00,0x33,0x80,0x01,0x70,0x01,0x0c,0x01,0x03,0x01,0x00,0x07,
-		0x0c,0x03,0x0d,0x01,0x01,0x01,0x00,0x01,0x40,0x01,0xa0,0x01,0x21,0x01,0x22,0x01,
-		0x12,0x01,0x11,0x03,0x09,0x01,0x08,0x04,0x04,0x02,0x08,0x02,0x10,0x05,0x11,0x01,
-		0x0f,0x01,0x01,0x02,0x00,0x36,0x80,0x05,0x40,0x05,0x20,0x05,0x10,0x05,0x08,0x05,
-		0x04,0x05,0x02,0x01,0x03,0x01,0x02,0x02,0x01,0x06,0x02,0x02,0x04,0x02,0x08,0x05,
-		0x07,0x01,0x00,0x3d};
-
-	static unsigned char const idle1[] PROGMEM = {165,
-		0x00,0x3b,0x80,0x02,0x40,0x04,0x20,0x04,0x10,0x01,0x08,0x01,0x04,0x01,0x02,0x02,
-		0x04,0x01,0x18,0x01,0x60,0x01,0x80,0x01,0x00,0x2f,0x80,0x03,0x00,0x31,0x3c,0x01,
-		0xc2,0x01,0x01,0x02,0x02,0x02,0x04,0x02,0x02,0x01,0x01,0x01,0x00,0x10,0x60,0x02,
-		0x00,0x01,0x81,0x01,0x82,0x02,0x84,0x01,0x08,0x01,0x10,0x01,0x20,0x01,0x40,0x01,
-		0x80,0x01,0x00,0x04,0x80,0x04,0x40,0x04,0x20,0x04,0x10,0x04,0x08,0x05,0x04,0x05,
-		0x02,0x04,0x01,0x04,0x00,0x33,0x80,0x01,0x70,0x01,0x19,0x01,0x06,0x01,0x00,0x07,
-		0x18,0x03,0x1b,0x01,0x03,0x01,0x00,0x01,0x40,0x01,0xa0,0x01,0x22,0x01,0x24,0x01,
-		0x14,0x01,0x12,0x03,0x0b,0x01,0x08,0x04,0x05,0x02,0x09,0x02,0x10,0x05,0x11,0x01,
-		0x0f,0x01,0x01,0x02,0x00,0x36,0x80,0x05,0x40,0x05,0x20,0x05,0x10,0x05,0x08,0x05,
-		0x04,0x05,0x02,0x01,0x03,0x01,0x02,0x02,0x01,0x06,0x02,0x02,0x04,0x02,0x08,0x05,
-		0x07,0x01,0x00,0x3d};
-
-	static unsigned char const idle2[] PROGMEM = {169,
-		0x00,0x3a,0x80,0x01,0x40,0x02,0x20,0x04,0x10,0x05,0x08,0x02,0x04,0x03,0x08,0x01,
-		0x30,0x01,0x40,0x01,0x80,0x01,0x00,0x2e,0x80,0x03,0x00,0x31,0x18,0x01,0x64,0x01,
-		0x82,0x01,0x02,0x05,0x01,0x01,0x00,0x04,0x80,0x02,0x00,0x09,0x80,0x01,0x00,0x01,
-		0x30,0x02,0x00,0x01,0xc0,0x01,0xc1,0x02,0xc2,0x01,0x04,0x01,0x08,0x01,0x10,0x01,
-		0x20,0x01,0x40,0x01,0x80,0x01,0x00,0x03,0x80,0x04,0x40,0x04,0x20,0x04,0x10,0x04,
-		0x08,0x05,0x04,0x05,0x02,0x04,0x01,0x04,0x00,0x33,0xc0,0x01,0x38,0x01,0x04,0x01,
-		0x03,0x01,0x00,0x07,0x0c,0x03,0x0d,0x01,0x01,0x01,0x00,0x01,0x40,0x01,0xa0,0x01,
-		0x21,0x01,0x22,0x01,0x12,0x01,0x11,0x03,0x09,0x01,0x08,0x04,0x04,0x02,0x08,0x02,
-		0x10,0x05,0x11,0x01,0x0f,0x01,0x01,0x02,0x00,0x36,0x80,0x05,0x40,0x05,0x20,0x05,
-		0x10,0x05,0x08,0x05,0x04,0x05,0x02,0x01,0x03,0x01,0x02,0x02,0x01,0x06,0x02,0x02,
-		0x04,0x02,0x08,0x05,0x07,0x01,0x00,0x3d};
-
-	static unsigned char const idle3[] PROGMEM = {171,
-		0x00,0x3a,0x80,0x01,0x40,0x02,0x20,0x04,0x10,0x04,0x08,0x02,0x04,0x01,0x02,0x03,
-		0x04,0x01,0x38,0x01,0x40,0x01,0x80,0x01,0x00,0x2e,0x80,0x03,0x00,0x31,0x1c,0x01,
-		0xe2,0x01,0x01,0x02,0x02,0x04,0x01,0x01,0x00,0x04,0x80,0x02,0x00,0x09,0x80,0x01,
-		0x00,0x01,0x30,0x02,0x00,0x01,0xc0,0x01,0xc1,0x02,0xc2,0x01,0x04,0x01,0x08,0x01,
-		0x10,0x01,0x20,0x01,0x40,0x01,0x80,0x01,0x00,0x03,0x80,0x04,0x40,0x04,0x20,0x04,
-		0x10,0x04,0x08,0x05,0x04,0x05,0x02,0x04,0x01,0x04,0x00,0x33,0x80,0x01,0x70,0x01,
-		0x0c,0x01,0x03,0x01,0x00,0x07,0x0c,0x03,0x0d,0x01,0x01,0x01,0x00,0x01,0x40,0x01,
-		0xa0,0x01,0x21,0x01,0x22,0x01,0x12,0x01,0x11,0x03,0x09,0x01,0x08,0x04,0x04,0x02,
-		0x08,0x02,0x10,0x05,0x11,0x01,0x0f,0x01,0x01,0x02,0x00,0x36,0x80,0x05,0x40,0x05,
-		0x20,0x05,0x10,0x05,0x08,0x05,0x04,0x05,0x02,0x01,0x03,0x01,0x02,0x02,0x01,0x06,
-		0x02,0x02,0x04,0x02,0x08,0x05,0x07,0x01,0x00,0x3d};
-
-	static unsigned char const *idle_anim[IDLE_FRAMES] = { idle0, idle0, idle1, idle2, idle3 };
-
-	static uint8_t current_frame = 0;
-	current_frame = (current_frame < IDLE_FRAMES - 1) ? current_frame + 1 : 0;
-
-	decode_frame(idle_anim[current_frame]);
-}
-
-static void render_cat_prep(void) {
-	// Run-length encoded prep frame
-	static unsigned char const prep[] PROGMEM = {181,
-		0x00,0x33,0x80,0x02,0x00,0x05,0x80,0x01,0x40,0x02,0x20,0x04,0x10,0x04,0x08,0x01,
-		0x04,0x01,0x02,0x01,0x01,0x02,0x02,0x01,0x0c,0x01,0x30,0x01,0x40,0x01,0x80,0x01,
-		0x00,0x2e,0x80,0x03,0x00,0x31,0x1e,0x01,0xe1,0x01,0x00,0x02,0x01,0x02,0x02,0x02,
-		0x81,0x01,0x80,0x02,0x00,0x02,0x80,0x02,0x00,0x09,0x80,0x01,0x00,0x01,0x30,0x02,
-		0x00,0x02,0x01,0x01,0xe1,0x01,0x1a,0x01,0x06,0x01,0x09,0x01,0x31,0x01,0x35,0x01,
-		0x01,0x01,0x8a,0x01,0x7c,0x01,0x00,0x02,0x80,0x04,0x40,0x04,0x20,0x04,0x10,0x04,
-		0x08,0x05,0x04,0x05,0x02,0x04,0x01,0x04,0x00,0x33,0x80,0x01,0x70,0x01,0x0c,0x01,
-		0x03,0x01,0x00,0x02,0x18,0x01,0x06,0x01,0x05,0x01,0x98,0x01,0x99,0x01,0x84,0x01,
-		0xc3,0x01,0x7c,0x01,0x41,0x02,0x40,0x02,0x20,0x01,0x21,0x01,0x22,0x01,0x12,0x01,
-		0x11,0x03,0x09,0x01,0x08,0x04,0x04,0x06,0x02,0x03,0x01,0x04,0x00,0x36,0x80,0x05,
-		0x40,0x05,0x20,0x05,0x10,0x05,0x08,0x05,0x04,0x05,0x02,0x01,0x03,0x01,0x02,0x02,
-		0x01,0x06,0x00,0x47};
-
-	decode_frame(prep);
-}
-
-static void render_cat_tap(void) {
-	// Run-length encoded tap frames
-	static unsigned char const tap0[] PROGMEM = {221,
-		0x00,0x33,0x80,0x02,0x00,0x05,0x80,0x01,0x40,0x02,0x20,0x04,0x10,0x04,0x08,0x01,
-		0x04,0x01,0x02,0x01,0x01,0x02,0x02,0x01,0x0c,0x01,0x30,0x01,0x40,0x01,0x80,0x01,
-		0x00,0x07,0xf8,0x04,0x00,0x05,0x80,0x02,0x00,0x1c,0x80,0x03,0x00,0x31,0x1e,0x01,
-		0xe1,0x01,0x00,0x02,0x01,0x02,0x02,0x02,0x81,0x01,0x80,0x02,0x00,0x02,0x80,0x02,
-		0x00,0x09,0x80,0x01,0x00,0x01,0x30,0x02,0x00,0x02,0x01,0x02,0x02,0x01,0x04,0x01,
-		0x08,0x01,0x10,0x01,0x20,0x01,0x43,0x01,0x87,0x01,0x07,0x01,0x01,0x01,0x00,0x01,
-		0xb8,0x01,0xbc,0x01,0xbe,0x01,0x9f,0x01,0x5f,0x02,0x4f,0x01,0x4c,0x01,0x20,0x04,
-		0x10,0x04,0x08,0x05,0x04,0x05,0x02,0x04,0x01,0x04,0x00,0x33,0x80,0x01,0x70,0x01,
-		0x0c,0x01,0x03,0x01,0x00,0x02,0x18,0x01,0x06,0x01,0x05,0x01,0x98,0x01,0x99,0x01,
-		0x84,0x01,0x43,0x01,0x7c,0x01,0x41,0x02,0x40,0x02,0x20,0x01,0x21,0x01,0x22,0x01,
-		0x12,0x01,0x11,0x03,0x09,0x01,0x08,0x04,0x04,0x02,0x08,0x02,0x10,0x05,0x11,0x01,
-		0x0f,0x01,0x01,0x01,0x3d,0x01,0x7c,0x01,0xfc,0x05,0x3c,0x01,0x0c,0x01,0x00,0x2e,
-		0x80,0x05,0x40,0x05,0x20,0x05,0x10,0x05,0x08,0x05,0x04,0x05,0x02,0x01,0x03,0x01,
-		0x02,0x02,0x01,0x06,0x00,0x24,0x01,0x01,0x03,0x02,0x00,0x20};
-
-	static unsigned char const tap1[] PROGMEM = {181,
-		0x00,0x33,0x80,0x02,0x00,0x05,0x80,0x01,0x40,0x02,0x20,0x04,0x10,0x04,0x08,0x01,
-		0x04,0x01,0x02,0x01,0x01,0x02,0x02,0x01,0x0c,0x01,0x30,0x01,0x40,0x01,0x80,0x01,
-		0x00,0x2e,0x80,0x03,0x00,0x31,0x1e,0x01,0xe1,0x01,0x00,0x02,0x01,0x02,0x02,0x02,
-		0x01,0x01,0x00,0x04,0x80,0x02,0x00,0x09,0x80,0x01,0x00,0x01,0x30,0x02,0x00,0x02,
-		0x01,0x01,0xe1,0x01,0x1a,0x01,0x06,0x01,0x09,0x01,0x31,0x01,0x35,0x01,0x01,0x01,
-		0x8a,0x01,0x7c,0x01,0x00,0x02,0x80,0x04,0x40,0x04,0x20,0x04,0x10,0x04,0x08,0x05,
-		0x04,0x05,0x02,0x04,0x01,0x04,0x00,0x33,0x80,0x01,0x70,0x01,0x0c,0x01,0x03,0x01,
-		0x00,0x0a,0x01,0x02,0x00,0x01,0x40,0x01,0xa0,0x01,0x21,0x01,0x22,0x01,0x12,0x01,
-		0x11,0x03,0x09,0x01,0x08,0x04,0x04,0x06,0x02,0x03,0x01,0x04,0x00,0x36,0x80,0x05,
-		0x40,0x05,0x20,0x05,0x10,0x05,0x08,0x05,0x04,0x05,0x02,0x01,0x03,0x01,0x7a,0x02,
-		0x79,0x04,0x39,0x01,0x31,0x01,0x02,0x02,0x04,0x02,0x08,0x03,0x88,0x02,0x87,0x01,
-		0x80,0x01,0x00,0x3c};
-
-	static unsigned char const *tap_anim[TAP_FRAMES] = { tap0, tap1 };
-
-	static unsigned char current_frame = 0;
-	current_frame = (current_frame + 1) & 1;
-
-	decode_frame(tap_anim[current_frame]);
-}
-
-void render_bongocat(void) {
-
-    void animate_cat(void) {
-        if (get_current_wpm() <= IDLE_SPEED) {
-            render_cat_idle();
-        }
-        if (get_current_wpm() > IDLE_SPEED && get_current_wpm() < TAP_SPEED) {
-            render_cat_prep();
-        }
-        if (get_current_wpm() >= TAP_SPEED){
-            render_cat_tap();
-        }
-    }
-
-    if (get_current_wpm() != 0) {
-        oled_on(); // not essential but turns on animation OLED with any alpha keypress
-        if (timer_elapsed32(anim_timer) > FRAME_DURATION) {
-            anim_timer = timer_read32();
-            animate_cat();
-        }
-        anim_sleep = timer_read32();
-    } else {
-        if (timer_elapsed32(anim_sleep) > OLED_TIMEOUT) {
-            oled_off();
-        } else {
-            if (timer_elapsed32(anim_timer) > FRAME_DURATION) {
-                anim_timer = timer_read32();
-                animate_cat();
-            }
-        }
-    }
-}
-
-// Write WPM information without using sprintf() from <stdio.h>
-// as the library increases firmware size
-void render_wpm_count(void) {
-    char wpm_str[4];
-    uint8_t wpm = get_current_wpm();
-
-    wpm_str[3] = '\0';
-    wpm_str[2] = '0' + wpm % 10;
-    wpm_str[1] = '0' + (wpm /= 10) % 10;
-    wpm_str[0] = '0' + wpm / 10;
-
-    oled_write_P(PSTR("WPM:"), false);
-    oled_write(wpm_str, false);
-}
-
-void render_wpm_mode(void) {
-    render_bongocat();
-    oled_set_cursor(0, 0);
-    render_wpm_count();
-}
-
-/*
- * Keyboard information mode, inspired by Satisfaction75.
- */
-
-static const char PROGMEM shift[2][4] = {
-    {0x89, 0x8a, 0x20, 0x00},
-    {0xa9, 0xaa, 0x20, 0x00}
-};
-static const char PROGMEM shift_pressed[2][4] = {
-    {0x80, 0x81, 0x88, 0x00},
-    {0xa0, 0xa1, 0xa8, 0x00}
-};
-
-static const char PROGMEM ctrl[2][4] = {
-    {0x8b, 0x8c, 0x20, 0x00},
-    {0xab, 0xac, 0x20, 0x00}
-};
-static const char PROGMEM ctrl_pressed[2][4] = {
-    {0x82, 0x83, 0x88, 0x00},
-    {0xa2, 0xa3, 0xa8, 0x00}
-};
-
-static const char PROGMEM optn[2][4] = {
-    {0x8d, 0x8e, 0x20, 0x00},
-    {0xad, 0xae, 0x20, 0x00}
-};
-static const char PROGMEM optn_pressed[2][4] = {
-    {0x84, 0x85, 0x88, 0x00},
-    {0xa4, 0xa5, 0xa8, 0x00}
-};
-
-static const char PROGMEM cmd[2][4] = {
-    {0x8f, 0x90, 0x20, 0x00},
-    {0xaf, 0xb0, 0x20, 0x00}
-};
-static const char PROGMEM cmd_pressed[2][4] = {
-    {0x86, 0x87, 0x88, 0x00},
-    {0xa6, 0xa7, 0xa8, 0x00}
-};
-
-static const char PROGMEM layer_str[6] = {0xc0, 0xc1, 0xc2, 0xc3, 0xc4, 0x00};
-static const char PROGMEM layer_row_1[4][3] = {
-    {0xc5, 0xc6, 0x00},
-    {0xc7, 0xc8, 0x00},
-    {0xc9, 0xca, 0x00},
-    {0xcb, 0xcc, 0x00}
-};
-static const char PROGMEM layer_row_2[3] = {0xcd, 0xce, 0x00};
-
-static void render_symbol(int x, int y, int num_rows, int row_length, const char data[][row_length + 1]) {
-    for (int row = 0; row < num_rows; row++) {
-        oled_set_cursor(x, y + row);
-        oled_write_P(data[row], false);
-    }
-}
-
-void render_layer_num(void) {
-    oled_set_cursor(1, 0);
-    oled_write_P(layer_str, false);
-
-    uint8_t layer = get_highest_layer(layer_state);
-		oled_set_cursor(1 + 5, 0);
-		oled_write_P(layer_row_1[layer], false);
-		oled_set_cursor(1 + 5, 1);
-		oled_write_P(layer_row_2, false);
-}
-
-void render_modifiers(void) {
-    uint8_t modifiers = get_mods();
-
-    if (modifiers & MOD_MASK_SHIFT)
-        render_symbol(0, 2, 2, 3, shift_pressed);
-    else
-        render_symbol(0, 2, 2, 3, shift);
-
-    if (modifiers & MOD_MASK_CTRL)
-        render_symbol(3, 2, 2, 3, ctrl_pressed);
-    else
-        render_symbol(3, 2, 2, 3, ctrl);
-    
-    if (modifiers & MOD_MASK_ALT)
-        render_symbol(6, 2, 2, 3, optn_pressed);
-    else
-        render_symbol(6, 2, 2, 3, optn);
-    
-    if (modifiers & MOD_MASK_GUI)
-        render_symbol(9, 2, 2, 3, cmd_pressed);
-    else
-        render_symbol(9, 2, 2, 3, cmd);
-}
-
-void render_info_mode(void) {
-    render_layer_num();
-    render_modifiers();
-}
-
-/* Init and rendering calls */
-
-// Rotate OLED display to the correct orientation
-oled_rotation_t oled_init_user(oled_rotation_t rotation) { return OLED_ROTATION_180; };
-
-void oled_task_user(void) {
-    switch (oled_mode) {
-        default:
-        case OLED_INFO:
-            render_info_mode();
-            break;
-        case OLED_WPM:
-            render_wpm_mode();
-            break;
-    }
-}
